+++ CHANGELOG.md
@@ -13,3 +13,5 @@
     - `ActionResult` allows both data and BadRequest
       - Strongly-typed responses with support for other status codes
   - Add consumes/produces Swagger annotations to the controller
+  - SwaggerGen option to set the API title
+  - SwaggerUI option to remove the "Try it Out" button